// src-ui/core/ipc/generated/debug/DebugCallResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DebugCallOutcome } from "./DebugCallOutcome";
import type { DebugError } from "./DebugError";

export type DebugCallResult = { "kind": "ok", outcome: DebugCallOutcome, } | { "kind": "err", error: DebugError, };