// src-ui/core/ipc/generated/debug/DebugChannelId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 一条调用 / 收事件的通道。会当作 map 键用，所以要 `Eq + Hash`
 */
export type DebugChannelId = { "kind": "auto" } | { "kind": "internal" } | { "kind": "http", name: string, } | { "kind": "ws", name: string, };